[update] integrate-spotify #145
